Correct the script of cosine_similarity.
