+++ api/tests_v2/configs/elementwise.json
@@ -16,6 +16,7 @@
             "type": "Variable"
         }
     },
+    "atol": 1E-5,
     "repeat": 5000
 }, {
     "op": "elementwise",
@@ -35,6 +36,7 @@
             "type": "Variable"
         }
     },
+    "atol": 1E-5,
     "repeat": 5000
 }, {
     "op": "elementwise",
